+++ src/ashley/machina_extensions/forum/apps.py
@@ -1,0 +1,8 @@
+"""Ashley application"""
+from machina.apps.forum.apps import ForumAppConfig as BaseForumAppConfig
+
+
+class ForumAppConfig(BaseForumAppConfig):
+    """Configuration class for the forum app."""
+
+    name = "ashley.machina_extensions.forum"